refactor(leaderboard): extract date range logic into a new utility file for better code organization and reusability
